Fix linter errors2
