+++ Gemfile
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 source "https://rubygems.org" # rubocop:disable Style/StringLiterals
 gem 'rubocop', '>= 1.0', '< 2.0'
 # gem "rails"
