+++ person.rb
@@ -1,0 +1,21 @@
+# frozen_string_literal: true
+
+# This is the Person class, parent of student and teacher class
+class Person
+  def initialize(age, id = 0, name = 'Unknown', parent_permission: true)
+    @id = id
+    @age = age
+    @name = name
+    @parent_permission = parent_permission
+  end
+  attr_reader :id
+  attr_accessor :age, :name
+
+  def can_use_services?
+    of_age? || @parent_permission
+  end
+
+  def of_age?
+    @age >= 18
+  end
+end
