+++ student.rb
@@ -1,5 +1,7 @@
-require_relative 'Person'
+# frozen_string_literal: true
 
+require_relative 'Person'
+# This is the Person class, parent of student and teacher class
 class Student < Person
   def initialize(age, classroom, name = 'Unknown', parent_permission: true)
     super(age: age, parent_permission: parent_permission, name: name)
